restructure component files
