+++ src/App.js
@@ -1,7 +1,7 @@
 import './App.css';
-import Header from './components/Header';
-import Expenses from './components/Expenses';
-import './components/Expenses.css';
+import Header from './components/UI/Header';
+import Expenses from './components/Expenses/Expenses';
+import './components/Expenses/Expenses.css';
 
 function App() {
   const expenses = [
