+++ src/components/Expenses/ExpenseItem.js
@@ -2,7 +2,7 @@ import React from 'react';
 import propTypes from 'prop-types';
 import './ExpenseItem.css';
 import ExpenseDate from './ExpenseDate';
-import Card from './Card';
+import Card from '../UI/Card';
 
 const ExpenseItem = (props) => {
   const {
